# feat: Allows setting wildcards for SkippedPropagatingNamespaces