**What type of PR is this?**
/kind feature

<!--
Add one of the following kinds:

/kind api-change
/kind bug
/kind cleanup
/kind deprecation
/kind design
/kind documentation
/kind failing-test
/kind feature
/kind flake

-->

**What this PR does / why we need it**:

Allows user settings to skip namespace propagation with specific prefixes/suffixes.

Like:   openshift-xxxxxx

<img width="348" alt="image" src="https://user-images.githubusercontent.com/17962021/230557127-bb560a8e-9d1e-4e55-ad52-6957b62a4b1a.png">







**Which issue(s) this PR fixes**:
Fixes #

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required.
-->
```release-note
`karmada-controller-manager`: Allows setting wildcards for SkippedPropagatingNamespaces.
```