@chaunceyjiang 2023-04-07
> Hi @chaunceyjiang, can you give some examples of how to use this parameter?

Example: 
```
 --skipped-propagating-namespaces=openshift-.*,example-namespace,.*-ns-suffixes
```