@RainbowMango 2023-04-13
I tested the `BenchmarkEventFilterNoMatchSkipNameSpaces` on my side and compared the performance changes before and after the patch(by [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat)), seems that this patch got 4.4% slower than before:
```
# benchstat BenchmarkEventFilterNoMatchSkipNameSpaces.before BenchmarkEventFilterNoMatchSkipNameSpaces.after 
goos: linux
goarch: amd64
pkg: github.com/karmada-io/karmada/pkg/detector
cpu: Intel(R) Xeon(R) Gold 6278C CPU @ 2.60GHz
                                   │ BenchmarkEventFilterNoMatchSkipNameSpaces.before │ BenchmarkEventFilterNoMatchSkipNameSpaces.after │
                                   │                      sec/op                      │          sec/op           vs base               │
EventFilterNoMatchSkipNameSpaces-4                                        1.199µ ± 1%                1.252µ ± 1%  +4.42% (p=0.000 n=15)
```
This seems acceptable, I'm surprised by the result though. I guess the reason might be we pre-compile the regular expression when parsing raw options at [here](https://github.com/karmada-io/karmada/pull/3373/files#diff-a111978a679a9f03e49c67a43ecb913299d153e59ce48d2d30f5029f27ab601aR565-R567).