@chaunceyjiang 2023-04-13
The compilation does have a large impact on the speed:

https://github.com/golang/go/issues/26623#issuecomment-408518236